+++ src/Symfony/Component/DependencyInjection/Compiler/AutowireRequiredMethodsPass.php
@@ -51,7 +51,7 @@ class AutowireRequiredMethodsPass extends AbstractRecursivePass
             while (true) {
                 if (false !== $doc = $r->getDocComment()) {
                     if (false !== stripos($doc, '@required') && preg_match('#(?:^/\*\*|\n\s*+\*)\s*+@required(?:\s|\*/$)#i', $doc)) {
-                        if (preg_match('#(?:^/\*\*|\n\s*+\*)\s*+@return\s++static[\s\*]#i', $doc)) {
+                        if ($this->isWither($reflectionMethod, $doc)) {
                             $withers[] = [$reflectionMethod->name, [], true];
                         } else {
                             $value->addMethodCall($reflectionMethod->name, []);
@@ -80,5 +80,21 @@ class AutowireRequiredMethodsPass extends AbstractRecursivePass
         }
 
         return $value;
+    }
+
+    private function isWither(\ReflectionMethod $reflectionMethod, string $doc): bool
+    {
+        $match = preg_match('#(?:^/\*\*|\n\s*+\*)\s*+@return\s++(static|\$this)[\s\*]#i', $doc, $matches);
+        if ($match && 'static' === $matches[1]) {
+            return true;
+        }
+
+        if ($match && '$this' === $matches[1]) {
+            return false;
+        }
+
+        $reflectionType = $reflectionMethod->hasReturnType() ? $reflectionMethod->getReturnType() : null;
+
+        return $reflectionType instanceof \ReflectionNamedType && 'static' === $reflectionType->getName();
     }
 }