+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -19,6 +19,7 @@ CHANGELOG
  * deprecated `Definition::getDeprecationMessage()`, use `Definition::getDeprecation()` instead
  * deprecated `Alias::getDeprecationMessage()`, use `Alias::getDeprecation()` instead
  * deprecated PHP-DSL's `inline()` function, use `service()` instead
+ * added support of PHP8 static return type for withers
 
 5.0.0
 -----